pyproject.toml
[tool.poetry]
name = "wikiparse"
version = "0.1.0"
description = "Wiktionary parsing for Finnish word definitions"
authors = ["Frankie Robertson <frankie@robertson.name>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.6"
sqlitedict = "*"
pandas = "*"
ujson = "*"
tblib = "*"
mediawiki-utilities = "*"
mwparserfromhell = "*"
boltons = "*"
click-log = "*"
mwxml = "*"
sqlalchemy = "*"
more-itertools = "*"
langdetect = "^1.0"
nltk = "^3.4"
wordfreq = "^2.2"
finntk = ">=0.0.68"
eralchemy = "^1.2.10"
hfst = "^3.15.0-beta.0"
typing_extensions = "^3.7.4"
snakemake = "^5.16.0"
orjson = "^3.0.0"
pytest = "^5.4.1"

[tool.poetry.dev-dependencies]
"nose2" = "*"
flake8 = "^3.5.0"
pre-commit = "^1.10.2"
mypy = "*"
black = "^19.10b0"

[tool.poetry.extras]
pgsql = ["psycopg2"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
